Update version 3.1
- Refactor creaete_or_update_unit will check if prevImgUrl has a data/url link or not, if has will delete from storage first then do the process.
- Refactor delete_unit will delete both unit and bill(has not been send) according to unit ID.
- Refactor upload_image will check for QRcode value from predict app, if there are none will return list of file name with corrupt QRcode through response, and the route will not create the data with corrupt QRcode(because we cannot identify which room the data is in).

+++ library/routes/unit_routes.py
@@ -3,9 +3,10 @@ from operator import and_
 from flask import request, jsonify, make_response
 import datetime as date
 from library.main import db, app, az
-from library.model.models import token_required, Unit
+from library.model.models import token_required, Unit, Bill
 from library.functions import toDate, pagination
 from dotenv import load_dotenv
+from library.functions import getNameFromURL
 
 load_dotenv()
 
@@ -29,11 +30,10 @@ def create_or_update_unit():
         if existing_unit:
 
             # Delete the oldest image url first
-            url = existing_unit.prevImgUrl
+            if existing_unit.prevImgUrl != '':
+                filename = getNameFromURL(existing_unit.prevImgUrl)
 
-            filename = url.rsplit('/', 1)[-1]
-
-            az.delete_file(filename)
+                az.delete_file(filename)
 
             # Then update the existing unit record
             existing_unit.prevNumberOfUnits = existing_unit.numberOfUnits
@@ -226,6 +226,10 @@ def delete_unit(current_user, role, rec_id):
     unit_record = Unit.query.filter_by(id=rec_id).first()
     if not unit_record:
         return make_response(jsonify({'message': 'Unit does not exist'}), 404)
+
+    find_related_bill = Bill.query.filter_by(unit_id=rec_id).first()
+    if find_related_bill:
+        db.session.delete(find_related_bill)
 
     url = unit_record.imgUrl
